requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.30.1
sqlmodel==0.0.14
aiosqlite==0.19.0
httpx==0.28.1
python-multipart==0.0.20
torch==2.4.1
torchvision==0.19.1
timm==1.0.22
tqdm==4.66.4
Pillow==10.4.0
reverse_geocoder==1.5.1
